Moved DBConnection String to Environmental Variable + Refactor
