+++ rueppellii-lvlup-asp.net-core/appsettings.json
@@ -1,8 +1,4 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=.;Database=lvlup;Trusted_Connection=True;MultipleActiveResultSets=true"
-  },
-
   "Logging": {
     "LogLevel": {
       "Default": "Warning"
